refactor: modularize candles feature
