+++ cmd/ingest/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"context"
 	"log"
-	"stock_backend/internal/feature/symbollist/adapters"
+	candlesadapters "stock_backend/internal/feature/candles/adapters"
+	symbollistadapters "stock_backend/internal/feature/symbollist/adapters"
 	"stock_backend/internal/infrastructure"
 	"stock_backend/internal/infrastructure/db"
-	"stock_backend/internal/infrastructure/mysql"
 	"stock_backend/internal/usecase"
 	"time"
 )
@@ -15,8 +15,8 @@ func main() {
 
 	db := db.OpenDB()
 	marketRepo := infrastructure.NewMarket()
-	candleRepo := mysql.NewCandleRepository(db)
-	symbolRepo := adapters.NewSymbolRepository(db)
+	candleRepo := candlesadapters.NewCandleRepository(db)
+	symbolRepo := symbollistadapters.NewSymbolRepository(db)
 	uc := usecase.NewIngestUsecase(marketRepo, candleRepo)
 
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
