+++ internal/domain/repository/market_repository.go
@@ -2,7 +2,7 @@ package repository
 
 import (
 	"context"
-	"stock_backend/internal/domain/entity"
+	"stock_backend/internal/feature/candles/domain/entity"
 )
 
 // MarketRepository　は株価データを取得するリポジトリのインターフェイスです。
