+++ internal/feature/candles/domain/entity/candle.go
