+++ internal/feature/candles/usecase/candles_usecase.go
@@ -2,8 +2,8 @@ package usecase
 
 import (
 	"context"
-	"stock_backend/internal/domain/entity"
-	"stock_backend/internal/domain/repository"
+	"stock_backend/internal/feature/candles/domain/entity"
+	"stock_backend/internal/feature/candles/domain/repository"
 )
 
 const (
