+++ internal/infrastructure/externalapi/twelvedata/market.go
@@ -7,8 +7,8 @@ import (
 	"log"
 	"net/http"
 	"net/url"
-	"stock_backend/internal/domain/entity"
 	"stock_backend/internal/domain/repository"
+	"stock_backend/internal/feature/candles/domain/entity"
 	"stock_backend/internal/interface/dto"
 	"strconv"
 	"time"
