+++ internal/usecase/ingest_usecase.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"log"
 	"stock_backend/internal/domain/repository"
+	candlerepo "stock_backend/internal/feature/candles/domain/repository"
 	"time"
 )
 
@@ -18,11 +19,11 @@ var ingestIntervals = []string{"1day", "1week", "1month"}
 // IngestUsecase は外部APIからデータを取得し、データベースに永続化するユースケースを定義します。
 type IngestUsecase struct {
 	market repository.MarketRepository
-	candle repository.CandleRepository
+	candle candlerepo.CandleRepository
 }
 
 // NewIngestUsecase は新しい IngestUsecase を作成します。
-func NewIngestUsecase(market repository.MarketRepository, candle repository.CandleRepository) *IngestUsecase {
+func NewIngestUsecase(market repository.MarketRepository, candle candlerepo.CandleRepository) *IngestUsecase {
 	return &IngestUsecase{market: market, candle: candle}
 }
